refactor(slots): extract routing.py (P3-slots §1i)
Moves the seeded slot catalogue + pure config-query routing helpers out
of slots/manager.py into slots/routing.py: SEEDED_SLOTS, NPU_SEEDED_SLOTS,
SLOT_ALIASES, LoadedSlot, _VALID_SLOT_TYPES, _SLOT_NAME_RE, seeded_slots,
default_slot_for, loaded_slot_from_config, loaded_slot, resolve_for_request,
route_for_request, add_slot, remove_slot.

Every function takes a narrow RoutingHost (iter_configs/_load_slot_config/
_resolve_alias/create/delete) instead of the concrete SlotManager, so it's
independently readable/testable; SlotManager keeps every name as a
one-line delegator and re-exports every moved symbol so
`from hal0.slots.manager import LoadedSlot` (tests, omni_router) and
`from hal0.slots.manager import SLOT_ALIASES` (dispatcher._capability_resolve,
api/__init__) stay unbroken.

tests/slots/test_manager.py: updated 4 `monkeypatch.setattr(...)` targets
from `hal0.slots.manager.shutil.which` to `hal0.slots.routing.shutil.which`
— the actual runtime shutil.which() call moved with seeded_slots().

No behavior change. tests/slots (683) + tests/stacks + model_meta +
omni_router green.
