+++ src/hal0/slots/routing.py
@@ -1,0 +1,391 @@
+"""Seeded slot catalogue + request-routing helpers (P3-slots §1i).
+
+Pure config-query routing: no state-machine coupling (never touches
+``_states``/``_transition``/locks). Extracted verbatim out of
+``hal0.slots.manager`` — every function here takes a narrow ``host``
+(anything exposing ``iter_configs``/``create``/``delete``, i.e. a
+:class:`hal0.slots.manager.SlotManager`) instead of assuming the full
+manager, so it stays independently readable. ``SlotManager`` keeps every
+name as a thin delegator (P3-slots §5 contract): ``seeded_slots``,
+``default_slot_for``, ``loaded_slot``, ``resolve_for_request``,
+``route_for_request``, ``add_slot``, ``remove_slot`` are all still callable
+exactly as before.
+
+Module-level re-exports from ``hal0.slots.manager`` (unchanged import
+paths): ``SEEDED_SLOTS``, ``NPU_SEEDED_SLOTS``, ``SLOT_ALIASES``,
+``LoadedSlot``.
+"""
+
+from __future__ import annotations
+
+import re
+import shutil
+from dataclasses import dataclass
+from typing import TYPE_CHECKING, Any, Protocol
+
+from hal0.model_meta import SLOT_TYPES
+from hal0.slots.state import SlotConfigError
+
+if TYPE_CHECKING:
+    from hal0.slots.manager import Slot
+
+# ── Seeded slot catalogue (PR-10, plan §4.2 + §10.2) ────────────────────────
+
+#: Slots that exist on every hal0 install regardless of hardware. The
+#: dashboard creates these as empty cards at first run; the bundle
+#: picker (Phase 5) populates their ``model.default`` fields. ``agent``
+#: is the GPU MoE chat-role sibling of ``chat`` (moved here from the NPU
+#: set in #679 — it is a GPU slot, not the NPU FLM anchor).
+# ADR-0023: `utility` (cheap helper) + `agent` (capable/default anchor) are the two
+# canonical llm seeds. `chat` is retired as a slot/role name (the `chat` *capability*
+# is unaffected; any llm slot serves it). `utility` is seeded so the memory
+# extraction target is always present on a fresh box.
+SEEDED_SLOTS: tuple[str, ...] = (
+    "utility",
+    "embed",
+    "rerank",
+    "stt",
+    "tts",
+    "img",
+    "vision",
+    "agent",
+)
+
+#: NPU FLM shadow slots seeded only when the FastFlowLM ``.deb`` is
+#: installed (``shutil.which('flm')`` truthy): the ASR + embed tags that
+#: ride the same coresident FLM process as the NPU chat anchor — the
+#: separate ``flm`` slot, NOT listed here. ``agent`` was previously
+#: (wrongly) in this set; it is a GPU chat-role slot and moved to
+#: SEEDED_SLOTS in #679. Opt-in at Pro+ bundle tier.
+#:
+#: Named ``{anchor}-stt`` / ``{anchor}-embed`` (anchor is ``flm``) to match
+#: the occupancy-pane virtual sub-cards (:mod:`hal0.api.routes.npu` synthesises
+#: ``s.name + "-stt"``) and the ``_touch_npu_shadow_count`` activity counters
+#: (:mod:`hal0.api.routes.v1`), which both key off that convention. The legacy
+#: ``stt-npu`` / ``embed-npu`` names never matched either surface and are
+#: migrated forward by :meth:`hal0.slots.npu.trio.reconcile_trio_slots`.
+NPU_SEEDED_SLOTS: tuple[str, ...] = ("flm-stt", "flm-embed")
+
+#: Back-compat alias map: old slot names → canonical new names.
+#: Aliases resolve transparently for dispatch and config lookup but are
+#: NEVER stored on disk and NEVER appear in list() / iter_configs() /
+#: /api/slots. ``agent-hermes`` maps to ``agent`` (a GPU seed slot, #679)
+#: so no new TOML is created — the alias just redirects old references.
+#: ADR-0023 retired the `primary` and `chat` aliases. The canonical roles are
+#: `agent` (default anchor) + `utility` (helper); a lingering operator-custom `chat`
+#: slot is reachable by its own name via generalized `hal0/<slot>` resolution, not an
+#: alias. Only the Hermes-era `agent-hermes` → `agent` redirect remains.
+SLOT_ALIASES: dict[str, str] = {
+    "agent-hermes": "agent",
+    # Route curated NPU model IDs to the FLM trio slot.
+    "qwen3:4b": "flm",
+    "qwen3-4b": "flm",
+}
+
+#: Slot ``type`` vocabulary (plan §4.1) — sourced from the canonical
+#: taxonomy (:data:`hal0.model_meta.SLOT_TYPES`) so validation, profiles,
+#: and /api/meta/enums can never drift.
+_VALID_SLOT_TYPES: frozenset[str] = frozenset(SLOT_TYPES)
+
+#: Slot-name policy: kebab-case, max 32 chars, leading alphanumeric.
+_SLOT_NAME_RE = re.compile(r"^[a-z0-9][a-z0-9_-]{0,31}$")
+
+
+@dataclass(frozen=True, slots=True)
+class LoadedSlot:
+    """Typed routing result for an enabled slot.
+
+    Returned by :meth:`SlotManager.resolve_for_request` and
+    :meth:`SlotManager.loaded_slot` so callers do not have to route to a
+    bare name and then reopen raw slot TOML to recover the model id,
+    device, labels, or system prompt.
+    """
+
+    name: str
+    model_id: str
+    slot_type: str
+    device: str
+    enabled: bool
+    labels: frozenset[str]
+    system_prompt: str = ""
+    profile: str | None = None
+    default: bool = False
+
+
+def seeded_slots(*, include_npu: bool | None = None) -> tuple[str, ...]:
+    """Return the seeded slot list, optionally including the NPU trio.
+
+    :data:`SEEDED_SLOTS` lands on every hal0 install. The NPU trio
+    shadows (``flm-stt`` / ``flm-embed``) only seed when the
+    FastFlowLM ``.deb`` is installed (``shutil.which('flm')``
+    truthy). Per plan §10.2 + §4.2.
+
+    Args:
+        include_npu: ``None`` (default) detects FLM presence at
+            runtime; ``True`` forces inclusion (tests + the bundle
+            picker's preview mode); ``False`` forces exclusion.
+    """
+    if include_npu is None:
+        include_npu = bool(shutil.which("flm"))
+    if include_npu:
+        return SEEDED_SLOTS + NPU_SEEDED_SLOTS
+    return SEEDED_SLOTS
+
+
+def loaded_slot_from_config(cfg: dict[str, Any]) -> LoadedSlot | None:
+    """Convert one raw slot config dict into a :class:`LoadedSlot`.
+
+    Returns ``None`` when the config does not describe an enabled slot
+    with a model id. The raw TOML shapes are intentionally absorbed here
+    so request routers and tool dispatchers consume a typed result.
+    """
+    name = str(cfg.get("name") or "").strip()
+    slot_type = str(cfg.get("type") or "").strip()
+    if not name or not slot_type:
+        return None
+    if cfg.get("enabled") is False:
+        return None
+
+    model_section = cfg.get("model") or {}
+    model_id = ""
+    if isinstance(model_section, dict):
+        raw_model = model_section.get("default", "")
+        if isinstance(raw_model, str):
+            model_id = raw_model.strip()
+    if not model_id:
+        return None
+
+    from hal0.model_meta import labels_of
+
+    raw_prompt = cfg.get("system_prompt")
+    system_prompt = raw_prompt if isinstance(raw_prompt, str) else ""
+    if not system_prompt:
+        extra = cfg.get("extra")
+        if isinstance(extra, dict) and isinstance(extra.get("system_prompt"), str):
+            system_prompt = extra["system_prompt"]
+
+    profile = cfg.get("profile")
+    return LoadedSlot(
+        name=name,
+        model_id=model_id,
+        slot_type=slot_type,
+        device=str(cfg.get("device") or ""),
+        enabled=True,
+        labels=frozenset(labels_of(cfg)),
+        system_prompt=system_prompt,
+        profile=profile if isinstance(profile, str) and profile else None,
+        default=cfg.get("default") is True,
+    )
+
+
+class RoutingHost(Protocol):
+    """Narrow seam routing needs from :class:`hal0.slots.manager.SlotManager`."""
+
+    async def iter_configs(self) -> list[dict[str, Any]]: ...
+    async def _load_slot_config(self, slot_name: str) -> dict[str, Any]: ...
+    @staticmethod
+    def _resolve_alias(name: str) -> str: ...
+    async def create(self, slot_name: str, slot_cfg: dict[str, Any]) -> Slot: ...
+    async def delete(self, slot_name: str, *, force: bool = False) -> None: ...
+
+
+async def default_slot_for(host: RoutingHost, slot_type: str) -> str | None:
+    """Return the name of the slot with ``type=slot_type`` and ``default=true``.
+
+    Plan §4.4 step 1. Exactly one ``default = true`` per type is
+    allowed; two defaults raise :class:`SlotConfigError` so the
+    misconfiguration surfaces at the routing call site instead of
+    silently picking one. Returns ``None`` when no slot of the
+    type has ``default = true`` (the caller is expected to
+    fall-through to the first enabled slot — see
+    :func:`route_for_request`).
+    """
+    candidates: list[str] = []
+    for cfg in await host.iter_configs():
+        if cfg.get("type") != slot_type:
+            continue
+        if cfg.get("default") is True:
+            candidates.append(str(cfg.get("name", "")))
+    if len(candidates) > 1:
+        raise SlotConfigError(
+            f"slot type {slot_type!r} has multiple default=true slots: "
+            f"{candidates}; exactly one is allowed",
+            details={"type": slot_type, "candidates": candidates},
+        )
+    return candidates[0] if candidates else None
+
+
+async def loaded_slot(host: RoutingHost, name: str) -> LoadedSlot | None:
+    """Return a typed view of an enabled configured slot, or ``None``.
+
+    Resolves back-compat aliases transparently. This is a read-only
+    inventory helper; it does not probe runtime state.
+    """
+    resolved = host._resolve_alias(name)
+    try:
+        cfg = await host._load_slot_config(resolved)
+    except SlotConfigError:
+        return None
+    return loaded_slot_from_config(cfg)
+
+
+async def resolve_for_request(
+    host: RoutingHost,
+    slot_type: str,
+    *,
+    required_labels: tuple[str, ...] = (),
+) -> LoadedSlot | None:
+    """Resolve a request of type ``slot_type`` to a loaded slot.
+
+    Plan §4.4 four-step routing:
+
+      1. **Type match + default.** If a slot of ``type=slot_type``
+         carries ``default = true``, prefer it.
+      2. **Label filter overlay.** When ``required_labels`` is
+         non-empty, the chosen slot's model must advertise every
+         required label (sourced from the slot's
+         ``model.labels`` list). The default is dropped if it
+         can't satisfy the overlay.
+      3. **Fall-through.** Otherwise pick the first ``enabled =
+         true`` slot of ``slot_type`` in TOML declaration order
+         (still satisfying the label overlay if any).
+      4. ``None`` when nothing matches.
+    Returning :class:`LoadedSlot` keeps callers from reopening raw slot
+    configs to discover the model id, labels, device, or system prompt.
+    """
+
+    def _satisfies(slot: LoadedSlot) -> bool:
+        if not required_labels:
+            return True
+        return set(required_labels).issubset(slot.labels)
+
+    slots = [
+        slot
+        for cfg in await host.iter_configs()
+        if cfg.get("type") == slot_type
+        for slot in [loaded_slot_from_config(cfg)]
+        if slot is not None
+    ]
+
+    # Step 1+2: try the default first.
+    default_name = await default_slot_for(host, slot_type)
+    if default_name is not None:
+        default_slot = next((slot for slot in slots if slot.name == default_name), None)
+        if default_slot is not None and _satisfies(default_slot):
+            return default_slot
+
+    # Step 3: fall-through to first enabled + label-matching slot.
+    for slot in slots:
+        if not _satisfies(slot):
+            continue
+        return slot
+
+    return None
+
+
+async def route_for_request(
+    host: RoutingHost,
+    slot_type: str,
+    *,
+    required_labels: tuple[str, ...] = (),
+) -> str | None:
+    """Resolve a request of type ``slot_type`` to a concrete slot name.
+
+    Compatibility wrapper for callers that have not moved to
+    :func:`resolve_for_request` yet.
+    """
+    slot = await resolve_for_request(host, slot_type, required_labels=required_labels)
+    return slot.name if slot is not None else None
+
+
+async def add_slot(
+    host: RoutingHost,
+    name: str,
+    *,
+    type: str,
+    model: str,
+    device: str = "gpu-rocm",
+    port: int = 8081,
+) -> Slot:
+    """Programmatic ``hal0 slot add`` (plan §4.3).
+
+    Validates kebab-case name, rejects seeded-name collisions,
+    rejects unknown slot types. The SlotConfig schema requires a
+    port in the 8081-8099 range.
+
+    Args:
+        name: Kebab-case identifier; must not collide with a
+            seeded slot (``SEEDED_SLOTS`` plus ``NPU_SEEDED_SLOTS``,
+            independently of whether FLM is installed).
+        type: One of ``llm | embedding | reranking | transcription
+            | tts | image``.
+        model: Model id to load by default.
+        device: Hardware preference (``gpu-rocm | gpu-vulkan | cpu
+            | npu``); see ``map_backend_to_device``. Default
+            ``gpu-rocm`` matches Strix Halo seed semantics.
+        port: SlotConfig.port — the container's loopback port.
+    """
+    if not _SLOT_NAME_RE.match(name):
+        raise SlotConfigError(
+            f"slot name {name!r}: use lowercase alphanumeric, hyphens, underscores; "
+            f"start with alphanumeric; max 32 chars",
+            details={"slot": name},
+        )
+    # Reject collisions with ALL seeded slots (include the NPU trio
+    # regardless of FLM presence — the names are reserved).
+    reserved = set(SEEDED_SLOTS) | set(NPU_SEEDED_SLOTS)
+    if name in reserved:
+        raise SlotConfigError(
+            f"slot {name!r} collides with a seeded slot; pick a different name",
+            details={"slot": name, "reserved": sorted(reserved)},
+        )
+    if type not in _VALID_SLOT_TYPES:
+        raise SlotConfigError(
+            f"slot type {type!r} is not one of {sorted(_VALID_SLOT_TYPES)}",
+            details={"slot": name, "type": type},
+        )
+    cfg = {
+        "name": name,
+        "port": port,
+        "type": type,
+        "device": device,
+        "provider": "llama-server",
+        "enabled": True,
+        "model": {"default": model},
+    }
+    return await host.create(name, cfg)
+
+
+async def remove_slot(host: RoutingHost, name: str) -> None:
+    """Programmatic ``hal0 slot remove`` (plan §4.3).
+
+    Rejects seeded-slot names (use :meth:`SlotManager.unload` or
+    ``capabilities.toml`` to disable a seeded slot, not delete it).
+    No side effect on the underlying model files — they stay in
+    the registry.
+    """
+    name = host._resolve_alias(name)
+    reserved = set(SEEDED_SLOTS) | set(NPU_SEEDED_SLOTS)
+    if name in reserved:
+        raise SlotConfigError(
+            f"slot {name!r} is seeded; cannot remove (disable it via capabilities.toml)",
+            details={"slot": name, "reserved": sorted(reserved)},
+        )
+    await host.delete(name)
+
+
+__all__ = [
+    "NPU_SEEDED_SLOTS",
+    "SEEDED_SLOTS",
+    "SLOT_ALIASES",
+    "LoadedSlot",
+    "RoutingHost",
+    "add_slot",
+    "default_slot_for",
+    "loaded_slot",
+    "loaded_slot_from_config",
+    "remove_slot",
+    "resolve_for_request",
+    "route_for_request",
+    "seeded_slots",
+]
